requirements.txt
paho-mqtt==1.5.0
networkx==2.4
rhasspy-asr~=0.2.0
rhasspy-hermes~=0.6.0
vosk==0.3.27
